+++ src/Bridge/Pdo/PdoBridge.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MakinaCorpus\QueryBuilder\Bridge\Pdo;
+
+class PdoBridge extends PdoQueryBuilder
+{
+}